[triton]fix precision, working on triton anymask bwd and lse correction

+++ cornstarch/kernel/triton/attn.py
@@ -95,7 +95,7 @@ def _attn_any_mask_fwd_inner(
         if fp8_v:
             p = p.to(tl.float8e5)
         else:
-            p = p.to(tl.float16)
+            p = p.to(v.dtype)
         acc = tl.dot(p, v, acc)
         # update m_i and l_i
         m_i = m_ij
@@ -280,14 +280,16 @@ def _attn_any_mask_bwd_dkdv(dk, dv,  #
         do = tl.load(do_ptrs)
         # Compute dV.
         ppT = pT
-        ppT = ppT.to(tl.float16)
+        # ppT = ppT.to(tl.float16)
+        ppT = ppT.to(v.dtype)
         dv += tl.dot(ppT, do)
         # D (= delta) is pre-divided by ds_scale.
         Di = tl.load(D + offs_m)
         # Compute dP and dS.
         dpT = tl.dot(v, tl.trans(do)).to(tl.float32)
         dsT = pT * (dpT - Di[None, :])
-        dsT = dsT.to(tl.float16)
+        # dsT = dsT.to(tl.float16)
+        dsT = dsT.to(v.dtype)
         dk += tl.dot(dsT, tl.trans(qT))
         # Increment pointers.
         curr_m += step_m
